--- a/cpp/cpp_boggle.cc
+++ b/cpp/cpp_boggle.cc
@@ -133,6 +133,8 @@ PYBIND11_MODULE(cpp_boggle, m) {
       )
       .def("get_children", &SumNode::GetChildren, py::return_value_policy::reference)
       .def("score_with_forces", &SumNode::ScoreWithForces)
+      .def("is_equal", &SumNode::IsEqual)
+      .def("hash", &SumNode::Hash)
       .def("orderly_bound", &SumNode::OrderlyBound);
 
   py::class_<ChoiceNode>(m, "ChoiceNode")
@@ -145,6 +147,8 @@ PYBIND11_MODULE(cpp_boggle, m) {
           &ChoiceNode::GetChildForLetter,
           py::return_value_policy::reference
       )
+      .def("is_equal", &ChoiceNode::IsEqual)
+      .def("hash", &ChoiceNode::Hash)
       .def(
           "get_children", &ChoiceNode::GetChildren, py::return_value_policy::reference
       );