This has a pretty dramatic effect on memory usage for the initial tree:

- Arena nodes: 113,017,781 → 10,475,032
- Arena bytes: 1.94G → 320M

Total RAM allocation during breaking is down 50%, but this translates into literally zero improvement in breaking time. This is pretty surprising to me! I wonder if it's more of a win on Intel, or when you're running 100 threads and memory contention is more of an issue.