@danvk 2025-12-29
I experimented more with this using `ankerl::unordered_dense::set` (https://github.com/martinus/unordered_dense, see `dag-ankerl` branch).

It’s ~2x faster than std::unordered_set. So the tradeoff for DAG is a marginal slowdown in build time and _maybe_ breaking time for ~4x less RAM.

- 156513
    - build_tree_secs: 2.58 → 6.81s
    - total time: 75.21 → 85.05s
    - tree_bytes: 1.12G → 256M
    - total_bytes: 1.62G → 768M
    - MRSS: 1.83G → 1.47G
- 705707
    - total time: 687.52 → 672.52s.
    - build_tree_secs: 4.891 → 10.569
    - tree_bytes: 1.94G → 384M
    - total_bytes: 2.94G → 1.31G
    - MRSS: 2.94G → 1.98G

So the tree build time isn’t *that* much worse, the bigger problem is that all subsequent steps get a bit slower, I assume because of worse memory locality / predictability. It remains interesting to me that this isn’t a win! I would have thought that 4x memory savings would have led to performance improvements.